[UD-43] Change OnSelectedPlaylist to be abstract
