+++ src/Assets/Scripts/PlaylistUtils/UIToolkitPlaylistsSelector.cs
@@ -1,6 +1,11 @@
-﻿using Assets.Scripts.PlaylistUtils;
+﻿using Assets.Scripts;
+using Assets.Scripts.PlaylistUtils;
+using Assets.Utils.HelperClasses;
 using Packages.StyngrSDK.Runtime.Scripts.HelperClasses;
+using Styngr.DTO.Response.SubscriptionsAndBundles;
+using Styngr.Exceptions;
 using Styngr.Model.Radio;
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UIElements;
@@ -50,6 +55,42 @@ public class UIToolkitPlaylistsSelector : PlaylistsSelector
 
         RegisterEvents();
 
+    }
+
+    protected override void OnPlaylistSelected(object sender, Playlist playlistInfo)
+    {
+        void OnSuccess(ActiveSubscription subscription)
+        {
+            playlistSelected?.Invoke(this, playlistInfo);
+            gameObject.SetActive(false);
+        }
+
+        void onFail(ErrorInfo error)
+        {
+            if (subscriptionHelper.IsPlaylistPremium(playlistInfo))
+            {
+                if (SubscriptionHelper.Instance.IsSubscriptionExpired(error.errorCode))
+                {
+                    InfoDialog.Instance.ShowErrorMessage("No Active Subscription", $"{error.errorMessage}.{Environment.NewLine}Please purchase a subscription or choose another playlist.");
+                    return;
+                }
+            }
+            else
+            {
+                playlistSelected?.Invoke(this, playlistInfo);
+                gameObject.SetActive(false);
+            }
+            Debug.LogError(error.Errors);
+        }
+
+        if (subscriptionManager != null)
+        {
+            subscriptionManager.GetActiveUserSubscription(OnSuccess, onFail);
+        }
+        else
+        {
+            OnSuccess(default);
+        }
     }
 
     private void OnPlaylistSelected(IEnumerable<object> enumerable)
